Norm all

+++ srcs/player/setup.c
@@ -6,7 +6,7 @@
 /*   By: tokino <tokino@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 14:31:25 by tokino            #+#    #+#             */
-/*   Updated: 2021/11/14 17:45:06 by tokino           ###   ########.fr       */
+/*   Updated: 2021/11/14 18:48:50 by tokino           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,8 +39,8 @@ void	set_player_position(t_player *player, t_map *map)
 			if (ft_strchr("NSEW", map->grid[row][col]))
 			{
 				found_player = true;
-				player->x = (float)(col + 0.5) * map->tile_size;
-				player->y = (float)(row + 0.5) * map->tile_size;
+				player->x = (col + 0.5) * map->tile_size;
+				player->y = (row + 0.5) * map->tile_size;
 				set_player_orientation(player, map->grid[row][col]);
 				map->grid[row][col] = '0';
 			}
